Point participants to the onboarding repo.
Part of the preparatory homework is to read the Carpentry
contribution guidelines. However, the link provided is to the
lesson template.

In this example, I think that this could cause confusion, because
the instructions talk about contributing to the current repo -- but
it leads to a page with FIXME links.